refactor(web): extract useClickOutside hook from GameCombobox
The reviewer asked why GameCombobox attaches mousedown/keydown listeners
on `document` instead of going through React. Short answer: there's no
React-native primitive for "user clicked outside this subtree" because
synthetic-event delegation only sees clicks rooted in the component
tree. Backdrop overlays solve the same problem at the cost of a
stacking-context layer; portal-based libraries like Radix do this in
their own internals.

Pull the listener wiring out of the component and into a small reusable
`useClickOutside(ref, onClose, enabled)` hook so the rationale lives in
one place (TSDoc on the hook) and any future popover-style component
can reuse it. The hook still attaches to `document` — that's the part
that fundamentally has to be document-level — but consumers now hand
in just a ref + callback. The effect is a no-op when `enabled` is
false so `useClickOutside(ref, close, open)` correctly strips listeners
while the popover is closed.

Reported by CoderCoco on #116.
